# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Internet Bank Backend API
  version: 1.1.0
  description: |
    Актуальная HTTP-спецификация backend-сервисов интернет-банка через API Gateway.
    Все запросы отправляются на один базовый URL.
servers:
  - url: http://localhost:8080
    description: API Gateway
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AppSettings:
      type: object
      required: [app_type, theme, hidden_account_ids]
      properties:
        app_type:
          type: string
          enum: [client, employee]
        theme:
          type: string
          enum: [light, dark]
        hidden_account_ids:
          type: array
          items:
            type: string
            format: uuid
...
    PutAppSettingsRequest:
      type: object
      required: [theme, hidden_account_ids]
      properties:
        theme:
          type: string
          enum: [light, dark]
        hidden_account_ids:
          type: array
          items:
            type: string
            format: uuid
...
paths:
...
  /app-settings/{appType}:
    get:
      tags: [AppSettings]
      summary: Получить настройки приложения текущего пользователя
      parameters:
        - in: path
          name: appType
          required: true
          schema:
            type: string
            enum: [client, employee]
      responses:
        '200':
          description: Настройки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSettings'
    put:
      tags: [AppSettings]
      summary: Обновить настройки приложения текущего пользователя
      parameters:
        - in: path
          name: appType
          required: true
          schema:
            type: string
            enum: [client, employee]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutAppSettingsRequest'
      responses:
        '200':
          description: Настройки обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSettings'